Added View for count of weapon bullets, logic of slider value changing separated. Some bugs fixed

+++ Assets/Source/Runtime/View/EnemyWavesSystem/IWavesCycleProgressView.cs
@@ -2,6 +2,6 @@ namespace SwampAttack.Runtime.View.EnemyWavesSystem
 {
     public interface IWavesCycleProgressView
     {
-        void Visualize(float completed, float maxValue);
+        void Visualize(int completed, int maxValue);
     }
 }
